Support zope.configuration 4.2+ and recent persistent
Use an 'OptionalDottedName' field for our ZCML 'id' type, because
that's what we really want, instead of `PythonIdentifier`.

If this still causes problems in the wild, this could go all the way
back to just a TextLine, but we really want the dots to enforce the
tree.

Also add 3.7 and drop 3.3 while here.

Tests will fail until zope.configuration 4.2.1 is released with
https://github.com/zopefoundation/zope.configuration/pull/37